test(enforce): one file per acceptance with binding header for #1424
Split the three grouped test files into nine per-acceptance files, each
carrying a '# Acceptance: acc:...' header in its leading comment block so the
tester bidirectional-binding validator anchors every acceptance to its test.
The header scan reads only the first 30 lines, which grouped files could not
satisfy for their 2nd/3rd acceptances.

validator-binding-must-be-bidirectional + phase-aware binding now clean;
full tester validator suite green (204 passed).
